create a library and executable
