+++ CMakeLists.txt
@@ -15,10 +15,10 @@ find_package(viam-cpp-sdk REQUIRED)
 # The image encoders
 set(THIRD_PARTY_SOURCES src/third_party/fpng.cpp src/third_party/lodepng.cpp)
 
-# construct the executable
-add_executable(viam-camera-realsense src/camera_realsense.cpp ${THIRD_PARTY_SOURCES})
+# construct the library
+add_library(viamrealsense src/camera_realsense.cpp ${THIRD_PARTY_SOURCES})
 
-target_link_libraries(viam-camera-realsense PRIVATE
+target_link_libraries(viamrealsense PRIVATE
     viam-cpp-sdk::viamsdk
     realsense2::realsense2
     turbojpeg
@@ -26,9 +26,26 @@ target_link_libraries(viam-camera-realsense PRIVATE
 )
 
 # Add all the flags
-target_compile_features(viam-camera-realsense PRIVATE 
+target_compile_features(viamrealsense PRIVATE 
     cxx_std_17
 )
+target_compile_options(viamrealsense PRIVATE 
+    -O2 
+    -Wredundant-move 
+    -Wpessimizing-move 
+)
+
+# construct the module executable
+add_executable(viam-camera-realsense main.cpp)
+
+target_link_libraries(viam-camera-realsense PRIVATE
+    viamrealsense
+    viam-cpp-sdk::viamsdk
+    realsense2::realsense2
+    turbojpeg
+    ${CMAKE_DL_LIBS}
+)
+
 target_compile_options(viam-camera-realsense PRIVATE 
     -O2 
     -Wredundant-move
